+++ test/github-issues/3576/entity/User.ts
@@ -1,0 +1,14 @@
+import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from "../../../../src";
+import { Book } from "./Book";
+
+@Entity()
+export class User {
+    @PrimaryGeneratedColumn("uuid")
+    id: string;
+
+    @Column()
+    name: string;
+
+    @OneToMany(type => Book, book => book.author, { lazy: true })
+    books: Book[];
+}